consensus/dccs: separation of data and logic for each hardfork
