+++ consensus/dccs/api.go
@@ -43,7 +43,7 @@ func (api *API) GetSnapshot(number *rpc.BlockNumber) (*Snapshot, error) {
 	if header == nil {
 		return nil, errUnknownBlock
 	}
-	return api.dccs.snapshot2(api.chain, header.Number.Uint64(), header.Hash(), nil)
+	return api.dccs.snapshot1(api.chain, header.Number.Uint64(), header.Hash(), nil)
 }
 
 // GetSnapshotAtHash retrieves the state snapshot at a given block.
@@ -52,7 +52,7 @@ func (api *API) GetSnapshotAtHash(hash common.Hash) (*Snapshot, error) {
 	if header == nil {
 		return nil, errUnknownBlock
 	}
-	return api.dccs.snapshot2(api.chain, header.Number.Uint64(), header.Hash(), nil)
+	return api.dccs.snapshot1(api.chain, header.Number.Uint64(), header.Hash(), nil)
 }
 
 // GetSigners retrieves the list of authorized signers at the specified block.
@@ -68,11 +68,11 @@ func (api *API) GetSigners(number *rpc.BlockNumber) ([]Signer, error) {
 	if header == nil {
 		return nil, errUnknownBlock
 	}
-	snap, err := api.dccs.snapshot2(api.chain, header.Number.Uint64(), header.Hash(), nil)
+	snap, err := api.dccs.snapshot1(api.chain, header.Number.Uint64(), header.Hash(), nil)
 	if err != nil {
 		return nil, err
 	}
-	return snap.signers2(), nil
+	return snap.signers1(), nil
 }
 
 // GetSignersAtHash retrieves the list of authorized signers at the specified block.
@@ -81,11 +81,11 @@ func (api *API) GetSignersAtHash(hash common.Hash) ([]Signer, error) {
 	if header == nil {
 		return nil, errUnknownBlock
 	}
-	snap, err := api.dccs.snapshot2(api.chain, header.Number.Uint64(), header.Hash(), nil)
+	snap, err := api.dccs.snapshot1(api.chain, header.Number.Uint64(), header.Hash(), nil)
 	if err != nil {
 		return nil, err
 	}
-	return snap.signers2(), nil
+	return snap.signers1(), nil
 }
 
 // Proposals returns the current proposals the node tries to uphold and vote on.
